[analyzer] avoid doing redundant work when resolving an extension override function reference.
The first thing done by
`FunctionReferenceResolver._resolvePropertyAccessFunction` (which is
responsible for resolving function references whose "function" part is
a property access) is to resolve the target of the property access. If
the target is an extension override, it then calls
`_resolveExtensionOverride` to finish resolution of the function
reference. Therefore, `_resolveExtensionOverride` doesn't need to
re-resolve the target; it's already been resolved.

This change helps pave the way for a follow-up CL in which I'll be
adding assertions that fire if the `ResolverVisitor` tries to resolve
a given subexpression multiple times. (Doing so is dangerous because
it can interfere with the correctness of flow analysis.)

Change-Id: Id2607b0325a78640c054f2978079d1f0afb35e67
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/400041
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Auto-Submit: Paul Berry <paulberry@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>