[analyzer] Fix "yield in non-generator" error reporting.
The analyzer unit tests `YieldEachInNonGeneratorTest.test_async` and
`YieldInNonGeneratorTest.test_async` were marked with a `@FailingTest`
annotation that claimed that the tests were failing because the parser
was trying to parse the yield statement as a binary expression. The
actual reason they were failing was because (a) both the analyzer and
the shared parser were redundantly flagging the error, resulting in
two errors rather than one, and (b) the tests had an incorrect offset
and length for the expected error.

As a result of these tests being marked with a `@FailingTest`
annotation, we had no test coverage of the
`YIELD_EACH_IN_NON_GENERATOR` and `YIELD_IN_NON_GENERATOR` errors.

I've resolved the redundancy by suppressing the error from the parser
and keeping the analyzer's logic to flag the error, since the
analyzer's logic produces a more specific error (it correctly
distinguishes between `YIELD_EACH_IN_NON_GENERATOR` and
`YIELD_IN_NON_GENERATOR`, whereas the shared parser does not). And
I've updated the tests with the correct offset and length for the
expected error. So the tests no longer fail and test coverage is
restored.

Change-Id: Ibbee3329d4699d7e78bec964303e9c9b239f0b8b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/400140
Auto-Submit: Paul Berry <paulberry@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>