[analyzer] Change ResolverVisitor to always visit subexpressions using analyzeExpression.
Previously, when `ResolverVisitor` needed to provide a context to a
subexpression, it would always use `analyzeExpression` to visit it,
but when no context was needed, it would sometimes visit the
subexpression by calling `ExpressionImpl.accept`. This made the
`ResolverVisitor` more difficult to reason about, since there wasn't a
common method through which all subexpression visits were dispatched.

With change, any time the `ResolverVisitor` visits an expression in a
context that's inside a method body or initializer, it visits it using
`analyzeExpression`. To make sure I haven't missed anything, and to
prevent future regressions, I've added an assertion to
`_InferenceLogWriterImpl.enterExpression` that verifies that all
expressions visited by the `ResolverVisitor` are visited via
`analyzeExpression` when inside a top level declaration.

This change paves the way for allowing the analyzer to use the shared
implementation of null-shorting introduced in
https://dart-review.googlesource.com/c/sdk/+/399480.

Change-Id: I56eb102ae0034898f5171ac4cc3e90676bce5db7
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/399781
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>