[analyzer] Fix flow analysis of null-aware extension overrides.
When an extension override is used in a null-aware fashion (for
example, `E(x)?.foo(...)`, where `E` refers to an extension
declaration), the expression appearing inside the extension override
(`x` in this example) needs to be passed to flow analysis, so that it
can be type promoted if appropriate.

Fixes https://github.com/dart-lang/sdk/issues/59654.

Bug: https://github.com/dart-lang/sdk/issues/59654
Change-Id: I9968f7403eee63f209eb39b12d4904151f7d6914
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/399623
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>