Plumb extra 'requireMain' flag through kernel builder entry point.
This is the method used by dart2wasm to generate a kernel from a source URI. For dynamic modules, the source will, by design, not include a 'main' function.

Without this flag the CFE will exit early with an error while compiling the source.

Change-Id: I1b66f7c0052ead80f5e555555d19a1b8a0cb590f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/400920
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Nate Biggs <natebiggs@google.com>