[analyzer] Use PromotableElementImpl2 to interface to shared analysis code.
Change the analyzer's use of shared classes and mixins to to supply
`PromotableElementImpl2` instead of `PromotableElement2` as the type
parameter that represents promotable variables.

This is a follow-up to
https://dart-review.googlesource.com/c/sdk/+/400660, which changed the
code from supplying `PromotableElement` to `PromotableElement2`. That
change was both larger and riskier than this one, since it changed the
identity of the objects that were passed to shared analysis code. This
change merely substitutes a more precise type for the type argument,
without changing the underlying objects in use.

This is part of a larger arc of work to change the analyzer's use of
the shared code so that the type parameters it supplies are not part
of the analyzer public API. See
https://github.com/dart-lang/sdk/issues/59763.

Change-Id: I2013dd0b7d38350ed7a9f2f4bd875d0006941754
Bug: https://github.com/dart-lang/sdk/issues/59763
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/401740
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>