[analyzer] Stop implementing SharedFunctionTypeStructure in analyzer public API.
When I introduced the `SharedFunctionTypeStructure` class in
https://dart-review.googlesource.com/c/sdk/+/386322, I failed to
realize that a side effect of this change was to expose the shared
methods `positionalParameterTypes`,
`requiredPositionalParameterCount`, and `sortedNamedParameters`
through the analyzer's public API.

To correct that mistake, I've moved the reference to
`SharedFunctionTypeStructure` from the `implements` clause of
`FunctionType` to the `implements` clause of `FunctionTypeImpl`.

To avoid this change causing a breakage for clients, I've also added
deprecated declarations of these three getters to the `FunctionType`
class. This ensures that if any analyzer clients have already started
depending on them, their code will continue to work, but they'll be
alerted to the fact that the members will be removed in the future.

This is part of a larger arc of work to change the analyzer's use of
the shared code so that the type parameters it supplies are not part
of the analyzer public API. See
https://github.com/dart-lang/sdk/issues/59763.

Bug: https://github.com/dart-lang/sdk/issues/59763
Change-Id: I442cbe29ed938ec2fed3a3fa65a95c9f0f47a38d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/401921
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>