[analyzer] Use TypeParameterElementImpl2 when interfacing with shared code.
Change the analyzer's use of the following shared generic types so
that it supplies the type parameter `TypeParameterElementImpl2`
instead of `TypeParameterElement` as the type it uses to represent
type parameters:

- `GeneratedTypeConstraint`
- `MergedTypeConstraint`
- `SharedFunctionTypeStructure`
- `SharedInferenceLogWriter`
- `SharedInferenceLogWriterImpl`
- `TypeAnalyzer`
- `TypeAnalyzerOperations`
- `TypeConstraintFromArgument`
- `TypeConstraintFromExtendsClause`
- `TypeConstraintFromFunctionContext`
- `TypeConstraintFromReturnType`
- `TypeConstraintGenerator`
- `TypeConstraintGeneratorMixin`
- `TypeConstraintOrigin`
- `UnknownTypeConstraintOrigin`

This required adding type casts in a few places. In other places, I
avoided type casts by tightening up some return types (for example,
`InstanceElementImpl2.typeParameters2` now returns
`List<TypeParameterElementImpl2>` instead of
`List<TypeParameterElement2>`. The analyzer public API was unchanged,
though (for example, `InstanceElement2.typeParameters2` still has a
return type of `List<TypeParameterElement2>`), so clients should be
unaffected.

As a result of this change, it is no longer necessary for a public API
class to implement the shared interface
`SharedTypeParameterStructure`. This interface was previously
implemented by `TypeParameterElement`; it is now implemented by
`TypeParameterElementImpl2`, which is not a part of the analyzer
public API.

This is part of a larger arc of work to change the analyzer's use of
the shared code so that the type parameters it supplies are not part
of the analyzer public API. See
https://github.com/dart-lang/sdk/issues/59763.

Change-Id: Ic6d42b415620430210e25ed212b11e03851d25ed
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/402500
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>