[analyzer] Use ExpressionImpl when interfacing with shared code
Change the analyzer's use of the following generic types so that it
supplies the type parameter `ExpressionImpl` instead of `Expression`
as the type of AST node it uses to represent expressions:

- `CaseHeadOrDefaultInfo`
- `FlowAnalysis`
- `MapPatternEntry`
- `MatchContext`
- `NullShortingMixin`
- `PropertyTarget`
- `SwitchExpressionMemberInfo`
- `SwitchStatementMemberInfo`
- `TypeAnalyzer`
- `TypeAnalyzerErrors`

This required adjusting some type casts and `is` tests to test against
concrete expression types instead of abstract public interface types
(e.g. `as SimpleIdentifierImpl` instead of `as SimpleIdentifier`).

In some places I was able to avoid adding casts by changing method
parameters to accept concrete statement types instead of abstract
public interface types. This required adding the `covariant` keyword
to some methods.

This is part of a larger arc of work to change the analyzer's use of
the shared code so that the type parameters it supplies are not part
of the analyzer public API. See
https://github.com/dart-lang/sdk/issues/59763.

Change-Id: Ib33f0f5b3b30bfe2554316b5aa7c11524b63bcd3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/403161
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>