[analyzer] Fix type of transformedParameters when inheriting covariance
In `InheritanceManager3._inheritCovariance`, if the original
`ExecutableElement` is in fact an `ExecutableMember`, then its
parameters will be `ParameterMember`s. These `ParameterMember`s need
to be converted to `ParameterElementImpl`s in order to be safely
stored in the synthetic `ExecutableElementImpl` that
`_inheritCovariance` returns.

I will follow up with a fix that changes the type accepted by
`ExecutableElementImpl.parameters=` in order to prevent similar
problems happening in the future.

Fixes https://github.com/dart-lang/sdk/issues/59849.

Bug: https://github.com/dart-lang/sdk/issues/59849
Change-Id: I636fd8cca0411745aedb3cd834354b4dcc0aadc1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/403322
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>