[dart2wasm] Add annotations to member and static intrinsics.
Each annotation corresponds to an enum value in the compiler that then dictates the generated code.

This is only done for static call and member intrinsics and not instance call intrinsics. This is because the target member of an instance intrinsic maybe be an interface rather than the annotated implementation of a method.

Change-Id: Iff8fe09b5078d81940cb1813fb65de50473519b1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/402960
Commit-Queue: Nate Biggs <natebiggs@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>