[analyzer] Use Impl types in TypeProviderImpl.
Change the methods in `TypeProviderImpl` so that their return types
are all "Impl" types rather than analyzer public API types.

This doesn't change the analyzer public API; it simply ensures that
clients of `TypeProviderImpl` within the analyzer itself won't have to
perform typecasts on the returned values in order to access the Impl
classes.

This is part of a larger arc of work to change the analyzer's use of
the shared code so that the type parameters it supplies are not part
of the analyzer public API (See
https://github.com/dart-lang/sdk/issues/59763). This change will
reduce the number of type casts that have to be done in order to pass
`TypeImpl` types to shared code rather than `DartType` types.

Change-Id: Ia351c3e8893cf10671b04b813ee80ff2b3bda341
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/402621
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>