[analyzer] Change DartType to implement SharedTypeStructure<TypeImpl>.
The type `DartType` is changed so that instead of implementing
`SharedTypeStructure<DartType>`, it implements
`SharedTypeStructure<TypeImpl>`.

This is part of a larger arc of work to change the analyzer's use of
the shared code so that the type parameters it supplies are not part
of the analyzer public API. See
https://github.com/dart-lang/sdk/issues/59763.

To ensure that this change doesn't expose the private `TypeImpl` type
through the analyzer public API, an abstract override of
`SharedTypeStructure.isStructurallyEqualTo` is added to `DartType`,
with a parameter type of `DartType`. (Without the override, the
parameter type would be `SharedTypeStructure<TypeImpl>`.) This method
has been marked as deprecated, because it was not intentional for it
to be exposed through the analyzer's public API.

The following types, derived from `DartType`, require similar changes
due to the fact that they implement shared types that are subtypes of
`SharedTypeStructure`:

- `DynamicTypeImpl` now implements
  `SharedDynamicTypeStructure<TypeImpl>`.

- `FunctionTypeImpl` now implements
  `SharedFunctionTypeStructure<TypeImpl, TypeParameterElementImpl2,
  FormalParameterElementOrMember>`.

- `InvalidTypeImpl` now implements
  `SharedInvalidTypeStructure<TypeImpl>`.

- `NullTypeImpl` now implements `SharedNullTypeStructure<TypeImpl>`.

- `RecordTypeImpl` now implements
  `SharedRecordTypeStructure<TypeImpl>`.

- `VoidTypeImpl` now implements `SharedVoidTypeStructure<TypeImpl>`.

- `UnknownInferredType` now implements
  `SharedUnknownTypeStructure<TypeImpl>`.

Since the type `SharedNamedFunctionParameterStructure` is tightly
integrated with `SharedFunctionTypeStructure`,
`FormalParameterElementOrMember` must also be changed, so that it
implements `SharedNamedFunctionParameterStructure<TypeImpl>` rather
than `SharedNamedFunctionParameterStructure<DartType>`.

Similarly, `TypeParameterElementImpl2` must be changed so that it
implements `SharedTypeParameterStructure<TypeImpl>` rather than
`SharedTypeParameterStructure<DartType>`.

Follow-up CLs will change other uses of shared generic types so that
they supply a type argument of `TypeImpl` rather than `DartType`.

Change-Id: Ib36491056d46e4cd706c0dc2b85adc5314cb0b2c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/403944
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>