[analyzer] Remove FormalParameterElementOrMember class.
The only reason this class was needed was to serve as a common
interface (not part of the analyzer public API) shared by
`ParameterMember` and `FormalParameterElementImpl`, which could be
used as a type argument when the analyzer instantiates generic classes
in `_fe_analyzer_shared`. Now that we have
`FormalParameterElementMixin`, it can serve as the common interface.

Change-Id: Idc25ca4077dc9c801d7eeb2e54b0d6b7809f1912
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/404362
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>